+++ packages/mint-components/src/components/sqm-portal-email-verification/PortalEmailVerification.stories.tsx
@@ -12,76 +12,70 @@ export default {
   },
 };
 
+const defaultCallbacks = {
+  submit: async (e) => await e,
+};
+
+const defaultContent = {
+  email: "email@example.com",
+  verifyMessage:
+    "A verification email was sent to {email}. Please verify your email to continue to the portal.",
+  emailVerificationHeader: "Verify your email",
+  resendEmailButtonText: "Re-send Email",
+  verificationStatusMessage: "Checking verification status in {countdown}",
+  verificationLoadingMessage: "Checking verification status",
+};
+
+const defaultStates = {
+  error: "",
+  loading: false,
+  success: false,
+  isVerified: false,
+  loadingVerification: false,
+  countdown: 10,
+};
+
 const defaultProps: PortalEmailVerificationViewProps = {
-  states: {
-    error: "",
-    loading: false,
-    success: false,
-  },
-  callbacks: {
-    submit: async (e) => await e,
-  },
-  content: {
-    email: "email@example.com",
-    verifyMessage:
-      "A verification email was sent to {email}. Please verify your email to continue to the portal.",
-    emailVerificationHeader: "Verify your email",
-    resendEmailButtonText: "Re-send Email",
-  },
+  states: defaultStates,
+  callbacks: defaultCallbacks,
+  content: defaultContent,
 };
 
 const errorProps: PortalEmailVerificationViewProps = {
   states: {
+    ...defaultStates,
     error: "Something went wrong. Please try again.",
-    loading: false,
-    success: false,
   },
-  callbacks: {
-    submit: async (e) => await e,
-  },
-  content: {
-    email: "email@example.com",
-    verifyMessage:
-      "A verification email was sent to {email}. Please verify your email to continue to the portal.",
-    emailVerificationHeader: "Verify your email",
-    resendEmailButtonText: "Re-send Email",
-  },
+  callbacks: defaultCallbacks,
+  content: defaultContent,
 };
 
 const loadingProps: PortalEmailVerificationViewProps = {
   states: {
-    error: "",
+    ...defaultStates,
     loading: true,
-    success: false,
   },
-  callbacks: {
-    submit: async (e) => await e,
-  },
-  content: {
-    email: "email@example.com",
-    verifyMessage:
-      "A verification email was sent to {email}. Please verify your email to continue to the portal.",
-    emailVerificationHeader: "Verify your email",
-    resendEmailButtonText: "Re-send Email",
-  },
+  callbacks: defaultCallbacks,
+  content: defaultContent,
 };
 
 const successProps: PortalEmailVerificationViewProps = {
   states: {
-    error: "",
-    loading: false,
+    ...defaultStates,
     success: true,
+    isVerified: true,
   },
-  callbacks: {
-    submit: async (e) => await e,
+  callbacks: defaultCallbacks,
+  content: defaultContent,
+};
+
+const loadingVerificationProps: PortalEmailVerificationViewProps = {
+  states: {
+    ...defaultStates,
+    loadingVerification: true,
   },
-  content: {
-    email: "email@example.com",
-    verifyMessage:
-      "A verification email was sent to {email}. Please verify your email to continue to the portal.",
-    emailVerificationHeader: "Verify your email",
-    resendEmailButtonText: "Re-send Email",
-  },
+  callbacks: defaultCallbacks,
+  content: defaultContent,
 };
 
 export const Default = () => <PortalEmailVerificationView {...defaultProps} />;
@@ -92,6 +86,10 @@ export const EmailVerificationWithError = () => (
 
 export const EmailVerificationLoading = () => (
   <PortalEmailVerificationView {...loadingProps} />
+);
+
+export const EmailVerificationCheckingVerification = () => (
+  <PortalEmailVerificationView {...loadingVerificationProps} />
 );
 
 export const EmailVerificationSuccess = () => (